requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
gradio>=4.0.0
Pillow>=9.0.0
numpy>=1.23.0
